--- a/go.mod
+++ b/go.mod
@@ -7,5 +7,6 @@ require (
 	github.com/btcsuite/btcd v0.20.1-beta
 	github.com/davecgh/go-spew v0.0.0-20171005155431-ecdeabc65495
 	github.com/kkdai/bstream v0.0.0-20161212061736-f391b8402d23
+	github.com/mr-tron/base58 v1.2.0
 	golang.org/x/crypto v0.0.0-20200115085410-6d4e4cb37c7d
 )