Adds an extra benchmark and dependency on external work by @mr-tron and myself, instead of copying code. Supersedes/closes https://github.com/btcsuite/btcutil/pull/93 

Provides significant time/space reduction in address encoding/decoding
    
- Before
```
    $ go test ./base58/ -bench=. -benchmem
    goos: darwin
    goarch: amd64
    pkg: github.com/btcsuite/btcutil/base58
    BenchmarkBase58EncodeAddr-8       195880              5864 ns/op           5.80 MB/s         648 B/op         51 allocs/op
    BenchmarkBase58DecodeAddr-8       290460              4115 ns/op          11.42 MB/s         264 B/op          8 allocs/op
    BenchmarkBase58Encode5k-8             18          63641483 ns/op           0.08 MB/s       82440 B/op       6725 allocs/op
    BenchmarkBase58Decode5k-8            226           5334819 ns/op           1.28 MB/s     1024873 B/op        380 allocs/op
```
 - After
```
    $ go test ./base58/ -bench=. -benchmem
    goos: darwin
    goarch: amd64
    pkg: github.com/btcsuite/btcutil/base58
    BenchmarkBase58EncodeAddr-8       857293              1362 ns/op          24.97 MB/s          96 B/op          2 allocs/op
    BenchmarkBase58DecodeAddr-8      2099786               566 ns/op          83.08 MB/s         128 B/op          2 allocs/op
    BenchmarkBase58Encode5k-8             45          26290144 ns/op           0.19 MB/s       13824 B/op          2 allocs/op
    BenchmarkBase58Decode5k-8            147           8173594 ns/op           0.84 MB/s       17152 B/op          2 allocs/op
```