@ribasushi 2020-06-16
Actually I had another insight, which brings the "after" to
```
pkg: github.com/btcsuite/btcutil/base58
BenchmarkBase58EncodeAddr-8   	 2690395	       436 ns/op	  77.94 MB/s	      96 B/op	       2 allocs/op
BenchmarkBase58DecodeAddr-8   	 3717728	       322 ns/op	 146.10 MB/s	     128 B/op	       2 allocs/op
BenchmarkBase58Encode5k-8     	     224	   5261993 ns/op	   0.95 MB/s	   13824 B/op	       2 allocs/op
BenchmarkBase58Decode5k-8     	     512	   2300409 ns/op	   2.97 MB/s	   15104 B/op	       2 allocs/op
```

I need to stabilize the new code a bit, and upstream it. Will open a new PR when ready.